+++ coordinate.h
@@ -23,7 +23,7 @@ typedef struct coordinate_t {
 #define toRAD *M_PI / 180
 #define toDEG *180 / M_PI
 
-void coordinate_init(coordinate_t *this, int64_t lat, int64_t lng, int64_t ele, int64_t timestamp);
+void coordinate_init(coordinate_t *this, double lat, double lng, int64_t ele, int64_t timestamp);
 void coordinate_deinit(coordinate_t *this);
 
 double get_distance(coordinate_t *point1, coordinate_t *point2);
